Fix crash opening the movement menu (64-bit pointer truncation)
Selecting a merc and clicking the movement menu (walk/run/sneak/crawl/
look/cancel) crashed with EXC_BAD_ACCESS in BtnMovementCallback.

GUI_BUTTON::UserData[4] was INT32. PopupMovementMenu stashed a UI_EVENT*
in UserData[0] via `(UINT32)(uintptr_t)pUIEvent`, truncating the 64-bit
pointer to 32 bits. The callback read it back as (UI_EVENT*)UserData[0]
-- a garbage pointer on 64-bit -- and crashed writing uiParams[1].

Widen UserData to uintptr_t (pointer-wide on 64-bit, still 32-bit on
Win32 so that build is unchanged) and drop the truncating (UINT32) cast
at the nine store sites. The other ~90 UserData uses only store small
ints, which are value-preserving in uintptr_t; nothing takes its address
as INT32* or depends on its byte size.

+++ sgp/Button System.h
@@ -104,7 +104,9 @@ typedef struct _GUI_BUTTON {
 	UINT32				uiOldFlags;				// Old flags from previous render loop
 	INT16					XLoc;							// Coordinates where button is on the screen
 	INT16					YLoc;
-	INT32					UserData[4];			// Place holder for user data etc.
+	uintptr_t				UserData[4];			// Place holder for user data etc. Pointer-wide
+																	// (was INT32) so callbacks can stash a pointer here
+																	// without truncating it on 64-bit; still 32-bit on Win32.
 	INT16					Group;						// Group this button belongs to (see DOCs)
 	INT8					bDefaultStatus;
 	//Button disabled style
